Moves Domain/Attribute and Domain/AttributeGroup ns to Product
