+++ src/Adapter/AttributeGroup/AttributeGroupViewDataProvider.php
@@ -27,8 +27,8 @@
 namespace PrestaShop\PrestaShop\Adapter\AttributeGroup;
 
 use AttributeGroup;
-use PrestaShop\PrestaShop\Core\Domain\AttributeGroup\Exception\AttributeGroupException;
-use PrestaShop\PrestaShop\Core\Domain\AttributeGroup\Exception\AttributeGroupNotFoundException;
+use PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Exception\AttributeGroupException;
+use PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Exception\AttributeGroupNotFoundException;
 use PrestaShopException;
 
 /**
