+++ src/PrestaShopBundle/Resources/config/services/adapter/attribute.yml
@@ -6,10 +6,10 @@ services:
     class: 'PrestaShop\PrestaShop\Adapter\Attribute\CommandHandler\DeleteAttributeHandler'
     tags:
       - name: 'tactician.handler'
-        command: 'PrestaShop\PrestaShop\Core\Domain\Attribute\Command\DeleteAttributeCommand'
+        command: 'PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Attribute\Command\DeleteAttributeCommand'
 
   prestashop.adapter.attribute.command_handler.bulk_delete_attribute_handler:
     class: 'PrestaShop\PrestaShop\Adapter\Attribute\CommandHandler\BulkDeleteAttributeHandler'
     tags:
       - name: 'tactician.handler'
-        command: 'PrestaShop\PrestaShop\Core\Domain\Attribute\Command\BulkDeleteAttributeCommand'
+        command: 'PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Attribute\Command\BulkDeleteAttributeCommand'
